refac: extract is desktop into util.js
